+++ Master/sample-acnode.ini
@@ -4,15 +4,16 @@ secret = ZeeSecret
 # Per node secrets; if none is specified; the above
 # default is used.
 #
-secrets = [ voordeur, machina, deurlezer=Othersecret, krachtstroom, deurleesmachien, deurbediener, ESP-dirkx, woodbandsawnode ]
+# secrets = [ voordeur, machina, deurlezer=Othersecret, krachtstroom, deurleesmachien, deurbediener, ESP-dirkx, woodbandsawnode ]
+secrets = [ test ]
 
 mqtthost = space.makerspaceleiden.nl
 topic = test
 
-smtphost = smtp.xs4all.nl
-alertsubject = '[AC-test]'
-alertfrom= spacelog@makerspaceleiden.nl
-alertto = unkown-tags@makerspaceleiden.nl
+#smtphost = smtp.xs4all.nl
+#alertsubject = '[AC-test]'
+#alertfrom= spacelog@makerspaceleiden.nl
+#alertto = unkown-tags@makerspaceleiden.nl
 
 dbfile = ./sample-keydb.txt
