+++ lib/DrumbeatNode.py
@@ -7,9 +7,10 @@ import hmac
 import hashlib
 
 sys.path.append('../lib')
-import ACNode as ACNode
-
+import ACNode
 class DrumbeatNode(ACNode.ACNode):
+# import SharedSecret
+# class DrumbeatNode(SharedSecret.SharedSecret):
   default_interval = 60
   default_node = "drumbeat"
 
@@ -18,19 +19,6 @@ class DrumbeatNode(ACNode.ACNode):
          help='DrumbeatNode interval, in seconds (default: '+str(self.default_interval)+' seconds)'),
 
     super().parseArguments()
-
-  def cmd_beat(self,path,node,theirbeat,payload):
-    if node != self.cnf.node:
-       self.logger.critical("Node {} sends a beat - yet this node {} is the beatmaster".format(node,self.cnf.node))
-    return
-
-  def cmd_announce(self,path,node,theirbeat,payload):
-    if node != self.cnf.node:
-       self.logger.info("Node '{}' (re)subscribed; sending beat.".format(node))
-       self.send(node,"beat")
-    else:
-       self.logger.info("Ignoring my own announce message.")
-    return None
 
   last_time = 0
   def loop(self):
